refactor: 関数をutilsファイルに分割
